# apps/local-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexura Local Server API",
    "version": "0.0.1",
    "description": "The per-branch local server's real API surface -- LAN-only, offline-first (Auth/Distribution doc Part 2: this server has no public URL; it's reachable only on the hotel's own network). Every path below was extracted directly from the real Express route definitions in `src/routes/*.ts` (see `scripts/generate-openapi.mjs`), not hand-typed -- run `npm run openapi:generate` after changing routes to keep this in sync. Auth is a JWT in an httpOnly `access_token` cookie (12h, with an 8h offline-continuation grace period -- see `src/auth/tokens.ts`), not a bearer token in the traditional OpenAPI sense; represented here as a cookie API key for tooling compatibility.",
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local dev server"
    },
    {
      "url": "http://nexura.local:4000",
      "description": "Real branch deployment (Auth doc Part 2, Layer 3) -- LAN hostname, no public DNS/TLS. `localhost` above is genuinely correct for dev, not a placeholder; this second entry is the real production shape."
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
  "paths": {
...
    "/settlements": {
      "get": {
        "operationId": "getSettlements",
        "tags": [
          "settlements"
        ],
        "summary": "GET /settlements",
        "description": "**Requires permission:** `payments:reconcile` or `finance:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postSettlements",
        "tags": [
          "settlements"
        ],
        "summary": "POST /settlements",
        "description": "**Requires permission:** `payments:reconcile`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/settlements/{id}/reconcile": {
      "post": {
        "operationId": "postSettlementsByIdReconcile",
        "tags": [
          "settlements"
        ],
        "summary": "POST /settlements/:id/reconcile",
        "description": "**Requires permission:** `payments:reconcile`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/settlements/report": {
      "get": {
        "operationId": "getSettlementsReport",
        "tags": [
          "settlements"
        ],
        "summary": "GET /settlements/report",
        "description": "**Requires permission:** `payments:reconcile` or `finance:read`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
...
  }
}